fix(Request Support): Url's without https not working
Fix an issue, where the link to the issue tracker would not work properly if entered without protocol. Move the parameter processing into a seperate function, to allow adding more required parameters and passing parameters to the fallback page (e.g. if the issue type is the only missing parameter, this allows to prefill all of the other inputs)

Close #93