feat(Request Support): Remove the "{addonName} is up to date" checkbox for all of the new error codes.
This commit removes the checkbox "{addonName} is up to date" for all error codes related to downloading/installation issues, since it's obvious that there is a newer version of the addon, so asking the user, whether the newest version of the addon is installed does not make sense. It does make sense for all of the other issues though, since newer versions might have made changes for the link to the endpoint url.

Closes #99