fix: :bug: Recursive Dirs not detecting addon.
Fix a bug with the logic of the RecursiveDirs class, where an addon wasn't found in a Zip directory, even though there was an addon. This was due to incorrect directory structures being created, when the subdirectories weren't listed seperately. Example: `["dir/", "dir/file.py"]` was working, but `["dir/file.py"]` wasn't.