fix(Updater): :bug: SAM Fails to check for updates.
When an enabled addon fails to register and thus doesn't appear in sys.modules, the update check fails because of a key error when trying to append the addon to the list of errors. This issue has been fixed by creating a function that checks, whether the module is in sys.modules before trying to access that module.

Fix #123